Rename grpc service
